Feat: Order

+++ core/src/Application/UseCase/Order/Delete.php
@@ -2,6 +2,8 @@
 
 namespace TechChallenge\Application\UseCase\Order;
 
+use TechChallenge\Domain\Order\Exceptions\OrderException;
+use TechChallenge\Domain\Order\Exceptions\OrderNotFoundException;
 use TechChallenge\Domain\Order\UseCase\Delete as IOrderUseCaseDelete;
 use TechChallenge\Domain\Order\UseCase\DtoInput;
 use TechChallenge\Domain\Order\Repository\IOrder as IOrderRepository;
@@ -14,7 +16,13 @@ class Delete implements IOrderUseCaseDelete
 
     public function execute(DtoInput $data): void
     {
-        $order = $this->OrderRepository->show($data->id);
+        if (is_null($data->getId()) || !$this->OrderRepository->exist(["id" => $data->getId()]))
+            throw new OrderNotFoundException();
+
+        $order = $this->OrderRepository->show(["id" => $data->getId()], true);
+
+        if (!$order->isNew())
+            throw new OrderException("Não pode excluir um pedido que não seja novo", 400);
 
         $order->delete();
